+++ ui/grom/lib/l10n/app_localizations.dart
@@ -370,6 +370,12 @@ abstract class AppLocalizations {
   /// **'Avg. speed'**
   String get workoutSpeedAvg;
 
+  /// No description provided for @workoutSpeedMax.
+  ///
+  /// In en, this message translates to:
+  /// **'Max. speed'**
+  String get workoutSpeedMax;
+
   /// No description provided for @workoutTotalTime.
   ///
   /// In en, this message translates to: