+++ ui/grom/lib/l10n/app_localizations_de.dart
@@ -150,6 +150,9 @@ class AppLocalizationsDe extends AppLocalizations {
   String get workoutSpeedAvg => 'Ø Geschwindigkeit';
 
   @override
+  String get workoutSpeedMax => 'Max. Geschwindigkeit';
+
+  @override
   String get workoutTotalTime => 'Gesamtzeit';
 
   @override