+++ ui/grom/lib/l10n/app_localizations_en.dart
@@ -149,6 +149,9 @@ class AppLocalizationsEn extends AppLocalizations {
   String get workoutSpeedAvg => 'Avg. speed';
 
   @override
+  String get workoutSpeedMax => 'Max. speed';
+
+  @override
   String get workoutTotalTime => 'Total time';
 
   @override