+++ ui/grom/lib/l10n/app_localizations_ru.dart
@@ -149,6 +149,9 @@ class AppLocalizationsRu extends AppLocalizations {
   String get workoutSpeedAvg => 'Средняя скорость';
 
   @override
+  String get workoutSpeedMax => 'Максимальная скорость';
+
+  @override
   String get workoutTotalTime => 'Общее время';
 
   @override